Made GetAllFileNames Generic for better scalability Made XMLReader classes specific for each BizoType
