+++ VEnitity/XML/Readers/GemCollectionXMLReader.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Xml;
+
+namespace VEntityFramework.XML
+{
+	internal class GemCollectionXMLReader : BaseXMLReader
+	{
+		protected override PropertyInfo GetPropertyFromXML(Type type, XmlNode child)
+		{
+			return child.Name == "Gem" 
+				? GetGem(type, child) 
+				: base.GetPropertyFromXML(type, child);
+		}
+
+		PropertyInfo GetGem(Type type, XmlNode node)
+		{
+			var name = "";
+			foreach (XmlNode child in node.ChildNodes)
+			{
+				if (child.Name == "Key")
+				{
+					name = child.InnerText.Replace(" ", "") + "Gem";
+				}
+			}
+			return type.GetProperty(name);
+		}
+	}
+}
