+++ VEnitity/XML/Readers/GemXMLReader.cs
@@ -1,0 +1,7 @@
+﻿namespace VEntityFramework.XML
+{
+	internal class GemXMLReader : BaseXMLReader 
+	{
+		protected override string XmlKey => "Key";
+	}
+}
