+++ VEnitity/XML/Readers/PerkXMLReader.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace VEntityFramework.XML
+{
+	internal class PerkXMLReader : BaseXMLReader
+	{
+		protected override string XmlKey => "Code";
+	}
+}
