+++ VEnitity/XML/Readers/SimpleXMLReader.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace VEntityFramework.XML
+{
+	internal class SimpleXMLReader : BaseXMLReader
+	{
+	}
+}
