+++ VEnitity/XML/Readers/SoulXMLReader.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace VEntityFramework.XML
+{
+	internal class SoulXMLReader : BaseXMLReader
+	{
+		protected override string XmlKey => "Type";
+	}
+}
