+++ VEnitity/XML/Readers/UnitConfigurationXMLReader.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace VEntityFramework.XML
+{
+	internal class UnitConfigurationXMLReader : BaseXMLReader
+	{
+		protected override string XmlKey => "Difficulty";
+	}
+}
